feat(combineLatest): add higher-order lettable version of combineLatest
